+++ src/bin/dune.ml
@@ -4,7 +4,7 @@ open Depgraph
 open Common
 open Std.Result_syntax
 
-let run path (`Type type_) (`Tred_modules tred_modules) (`Tred_libraries tred_libraries) (`With_pps with_pps) =
+let run path (`Type type_) (`Tred_modules tred_modules) (`Tred_libraries tred_libraries) (`With_modules with_modules) (`With_pps with_pps) =
   let f = Fpath.v path in
   let* dir =
     let+ is_dir = OS.Dir.exists f in
@@ -14,6 +14,7 @@ let run path (`Type type_) (`Tred_modules tred_modules) (`Tred_libraries tred_li
       Fpath.parent f
   in
   let+ s = OS.Dir.with_current dir (fun () ->
+      (* TODO: don't add --with-deps if with_modules = false *)
       let dune_describe = Cmd.(v "dune" % "describe" % "workspace" % "--with-deps") in
       match with_pps with
       | None ->
@@ -30,7 +31,7 @@ let run path (`Type type_) (`Tred_modules tred_modules) (`Tred_libraries tred_li
       ) ()
     |> Result.join
   in
-  Dune_describe_graph.g_of_string ~tred_modules ~tred_libraries s
+  Dune_describe_graph.g_of_string ~tred_modules ~tred_libraries ~with_modules s
   |> output type_
 
 open Cmdliner
@@ -44,7 +45,7 @@ let with_pps =
   Arg.(term_map (fun b -> `With_pps b) & value & opt (some bool) None & info ["with-pps"] ~docv:"BOOL" ~doc)
 
 let term =
-  Term.cli_parse_result Term.(const run $ path $ Common.type_ $ Common.tred_modules $ Common.tred_libraries $ with_pps)
+  Term.cli_parse_result Term.(const run $ path $ Common.type_ $ Common.tred_modules $ Common.tred_libraries $ Common.with_modules $ with_pps)
 
 let cmd =
   let doc = "Generate dependency graph from dune project." in