+++ src/bin/dune_describe_file.ml
@@ -3,9 +3,9 @@ open Bos
 open Depgraph
 open Std.Result_syntax
 
-let run file (`Type type_) (`Tred_modules tred_modules) (`Tred_libraries tred_libraries) =
+let run file (`Type type_) (`Tred_modules tred_modules) (`Tred_libraries tred_libraries) (`With_modules with_modules) =
   let+ s = OS.File.read (Fpath.v file) in
-  Dune_describe_graph.g_of_string ~tred_modules ~tred_libraries s
+  Dune_describe_graph.g_of_string ~tred_modules ~tred_libraries ~with_modules s
   |> output type_
 
 open Cmdliner
@@ -15,7 +15,7 @@ let file =
   Arg.(required & pos 0 (some non_dir_file) None & info [] ~docv:"FILE" ~doc)
 
 let term =
-  Term.cli_parse_result Term.(const run $ file $ Common.type_ $ Common.tred_modules $ Common.tred_libraries)
+  Term.cli_parse_result Term.(const run $ file $ Common.type_ $ Common.tred_modules $ Common.tred_libraries $ Common.with_modules)
 
 let cmd =
   let doc = Format.sprintf "Generate dependency graph from 'dune describe workspace' output." in